Re-work job status fetching
Main motivation is to simplify project structure and new workloads
onboarding. Second goal is to introduce fine-grained status check based
on both system level and workload level.

1. Split status it two: job status (runner level, like slurm job) and
   workload status (per-workload).
2. Runner level always returns OK replicating Default strategy.
3. Per-workload logic moved into `TestDefinition.was_run_successfull`.
