+++ src/cloudai/_core/test_template.py
@@ -19,8 +19,6 @@ from typing import Any, Dict, Optional
 
 from .command_gen_strategy import CommandGenStrategy
 from .grading_strategy import GradingStrategy
-from .job_status_result import JobStatusResult
-from .job_status_retrieval_strategy import JobStatusRetrievalStrategy
 from .json_gen_strategy import JsonGenStrategy
 from .system import System
 from .test_scenario import TestRun
@@ -39,7 +37,6 @@ class TestTemplate:
         command_gen_strategy (CommandGenStrategy): Strategy for generating execution commands.
         json_gen_strategy (JsonGenStrategy): Strategy for generating json string.
         grading_strategy (GradingStrategy): Strategy for grading performance based on test outcomes.
-        job_status_retrieval_strategy (JobStatusRetrievalStrategy): Strategy for determining job statuses.
     """
 
     __test__ = False
@@ -54,7 +51,6 @@ class TestTemplate:
         self.system = system
         self._command_gen_strategy: Optional[CommandGenStrategy] = None
         self._json_gen_strategy: Optional[JsonGenStrategy] = None
-        self.job_status_retrieval_strategy: Optional[JobStatusRetrievalStrategy] = None
         self.grading_strategy: Optional[GradingStrategy] = None
 
     @property
@@ -111,23 +107,6 @@ class TestTemplate:
                 "by calling the appropriate registration function for the system type."
             )
         return self.json_gen_strategy.gen_json(tr)
-
-    def get_job_status(self, output_path: Path) -> JobStatusResult:
-        """
-        Determine the job status by evaluating the contents or results in a specified output directory.
-
-        Args:
-            output_path (Path): Path to the output directory.
-
-        Returns:
-            JobStatusResult: The result containing the job status and an optional error message.
-        """
-        if self.job_status_retrieval_strategy is None:
-            raise ValueError(
-                "job_status_retrieval_strategy is missing. Ensure the strategy is registered in "
-                "the Registry by calling the appropriate registration function for the system type."
-            )
-        return self.job_status_retrieval_strategy.get_job_status(output_path)
 
     def grade(self, directory_path: Path, ideal_perf: float) -> Optional[float]:
         """
