+++ src/cloudai/test_parser.py
@@ -24,7 +24,6 @@ from pydantic import BaseModel, ValidationError
 from .core import (
     CommandGenStrategy,
     GradingStrategy,
-    JobStatusRetrievalStrategy,
     JsonGenStrategy,
     Registry,
     System,
@@ -116,28 +115,16 @@ class TestParser:
 
     def _fetch_strategy(  # noqa: D417
         self,
-        strategy_interface: Type[
-            Union[
-                TestTemplateStrategy,
-                JobStatusRetrievalStrategy,
-                GradingStrategy,
-            ]
-        ],
+        strategy_interface: Type[Union[TestTemplateStrategy, GradingStrategy]],
         system_type: Type[System],
         test_definition_type: Type[TestDefinition],
         cmd_args: Dict[str, Any],
-    ) -> Optional[
-        Union[
-            TestTemplateStrategy,
-            JobStatusRetrievalStrategy,
-            GradingStrategy,
-        ]
-    ]:
+    ) -> Optional[Union[TestTemplateStrategy, GradingStrategy]]:
         """
         Fetch a strategy from the registry based on system and template.
 
         Args:
-            strategy_interface (Type[Union[TestTemplateStrategy, JobStatusRetrievalStrategy]]):
+            strategy_interface (Type[Union[TestTemplateStrategy, GradingStrategy]]):
                 The strategy interface to fetch.
             system_type (Type[System]): The system type.
             test_template_type (Type[TestTemplate]): The test template type.
@@ -183,10 +170,6 @@ class TestParser:
         obj.json_gen_strategy = cast(
             JsonGenStrategy,
             self._fetch_strategy(JsonGenStrategy, type(obj.system), type(tdef), cmd_args),
-        )
-        obj.job_status_retrieval_strategy = cast(
-            JobStatusRetrievalStrategy,
-            self._fetch_strategy(JobStatusRetrievalStrategy, type(obj.system), type(tdef), cmd_args),
         )
         obj.grading_strategy = cast(
             GradingStrategy, self._fetch_strategy(GradingStrategy, type(obj.system), type(tdef), cmd_args)
