+++ src/cloudai/workloads/nccl_test/nccl.py
@@ -16,7 +16,7 @@
 
 from typing import Literal, Optional, Union
 
-from cloudai.core import DockerImage, Installable
+from cloudai.core import DockerImage, Installable, JobStatusResult, TestRun
 from cloudai.models.workload import CmdArgs, TestDefinition
 
 
@@ -121,3 +121,63 @@ class NCCLTestDefinition(TestDefinition):
     @property
     def installables(self) -> list[Installable]:
         return [self.docker_image, self.predictor] if self.predictor else [self.docker_image]
+
+    def was_run_successful(self, tr: TestRun) -> JobStatusResult:
+        stdout_path = tr.output_path / "stdout.txt"
+        if stdout_path.is_file():
+            with stdout_path.open("r") as file:
+                content = file.read()
+
+                # Check for specific error patterns
+                if "Test NCCL failure" in content:
+                    return JobStatusResult(
+                        is_successful=False,
+                        error_message=(
+                            f"NCCL test failure detected in {stdout_path}. "
+                            "Possible reasons include network errors or remote process exits. "
+                            "Please review the NCCL test output and errors in the file first. "
+                            "If the issue persists, contact the system administrator."
+                        ),
+                    )
+                if "Test failure" in content:
+                    return JobStatusResult(
+                        is_successful=False,
+                        error_message=(
+                            f"Test failure detected in {stdout_path}. "
+                            "Please review the specific test failure messages in the file. "
+                            "Ensure that the NCCL test environment is correctly set up and configured. "
+                            "If the issue persists, contact the system administrator."
+                        ),
+                    )
+
+                # Check for success indicators
+                if "# Out of bounds values" in content and "# Avg bus bandwidth" in content:
+                    return JobStatusResult(is_successful=True)
+
+                # Identify missing success indicators
+                missing_indicators = []
+                if "# Out of bounds values" not in content:
+                    missing_indicators.append("'# Out of bounds values'")
+                if "# Avg bus bandwidth" not in content:
+                    missing_indicators.append("'# Avg bus bandwidth'")
+
+                error_message = (
+                    f"Missing success indicators in {stdout_path}: {', '.join(missing_indicators)}. "
+                    "These keywords are expected to be present in stdout.txt, usually towards the end of the file. "
+                    "Please review the NCCL test output and errors in the file. "
+                    "Ensure the NCCL test ran to completion. You can run the generated sbatch script manually "
+                    f"and check if {stdout_path} is created and contains the expected keywords. "
+                    "If the issue persists, contact the system administrator."
+                )
+                return JobStatusResult(is_successful=False, error_message=error_message)
+
+        return JobStatusResult(
+            is_successful=False,
+            error_message=(
+                f"stdout.txt file not found in the specified output directory {tr.output_path}. "
+                "This file is expected to be created as a result of the NCCL test run. "
+                "Please ensure the NCCL test was executed properly and that stdout.txt is generated. "
+                f"You can run the generated NCCL test command manually and verify the creation of {stdout_path}. "
+                "If the issue persists, contact the system administrator."
+            ),
+        )
