+++ src/cloudai/workloads/nemo_run/__init__.py
@@ -15,7 +15,6 @@
 # limitations under the License.
 
 from .data_store_report_generation_strategy import NeMoRunDataStoreReportGenerationStrategy
-from .job_status_retrieval_strategy import NeMoRunJobStatusRetrievalStrategy
 from .nemo_run import Data, Log, LogCkpt, NeMoRunCmdArgs, NeMoRunTestDefinition, Trainer, TrainerStrategy
 from .report_generation_strategy import NeMoRunReportGenerationStrategy
 from .slurm_command_gen_strategy import NeMoRunSlurmCommandGenStrategy
@@ -26,7 +25,6 @@ __all__ = [
     "LogCkpt",
     "NeMoRunCmdArgs",
     "NeMoRunDataStoreReportGenerationStrategy",
-    "NeMoRunJobStatusRetrievalStrategy",
     "NeMoRunReportGenerationStrategy",
     "NeMoRunSlurmCommandGenStrategy",
     "NeMoRunTestDefinition",
