+++ src/cloudai/workloads/nixl_bench/__init__.py
@@ -14,7 +14,6 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-from .job_status_retrieval_strategy import NIXLBenchJobStatusRetrievalStrategy
 from .nixl_bench import NIXLBenchCmdArgs, NIXLBenchTestDefinition
 from .nixl_summary_report import NIXLBenchSummaryReport
 from .report_generation_strategy import NIXLBenchReportGenerationStrategy
@@ -22,7 +21,6 @@ from .slurm_command_gen_strategy import NIXLBenchSlurmCommandGenStrategy
 
 __all__ = [
     "NIXLBenchCmdArgs",
-    "NIXLBenchJobStatusRetrievalStrategy",
     "NIXLBenchReportGenerationStrategy",
     "NIXLBenchSlurmCommandGenStrategy",
     "NIXLBenchSummaryReport",
