+++ src/cloudai/workloads/nixl_bench/nixl_bench.py
@@ -16,7 +16,7 @@
 
 from typing import Optional
 
-from cloudai.core import DockerImage, Installable
+from cloudai.core import DockerImage, Installable, JobStatusResult, TestRun
 from cloudai.models.workload import CmdArgs, TestDefinition
 
 
@@ -51,3 +51,31 @@ class NIXLBenchTestDefinition(TestDefinition):
     @property
     def installables(self) -> list[Installable]:
         return [self.docker_image, *self.git_repos, self.etcd_image]
+
+    def was_run_successful(self, tr: TestRun) -> JobStatusResult:
+        stdout_path = tr.output_path / "stdout.txt"
+        if not stdout_path.exists():
+            return JobStatusResult(
+                is_successful=False,
+                error_message=f"stdout.txt file not found in the specified output directory {tr.output_path}.",
+            )
+
+        has_header, has_data = False, False
+        for line in stdout_path.read_text().splitlines():
+            if "Block Size (B)      Batch Size     Avg Lat. (us)  B/W (MiB/Sec)  B/W (GiB/Sec)  B/W (GB/Sec)" in line:
+                has_header = True
+                continue
+            if has_header and len(line.split()) == 6:
+                has_data = True
+                break
+
+        if has_data:
+            return JobStatusResult(is_successful=True)
+
+        if not has_header:
+            return JobStatusResult(
+                is_successful=False,
+                error_message=f"NIXLBench results table not found in {stdout_path}.",
+            )
+
+        return JobStatusResult(is_successful=False, error_message=f"NIXLBench data not found in {stdout_path}.")
