+++ tests/job_status_retrieval_strategy/test_jax_toolbox_job_status_retrieval_strategy.py
@@ -14,30 +14,36 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-from pathlib import Path
 
-from cloudai.workloads.jax_toolbox import JaxToolboxJobStatusRetrievalStrategy
+from cloudai.core import TestRun
+from cloudai.workloads.jax_toolbox import JaxToolboxCmdArgs, JaxToolboxTestDefinition
 
 
-class TestJaxToolboxJobStatusRetrievalStrategy:
+class TestJaxToolboxSuccessCheck:
     def setup_method(self) -> None:
-        self.js = JaxToolboxJobStatusRetrievalStrategy()
+        self.jax_tdef = JaxToolboxTestDefinition(
+            name="n",
+            description="d",
+            test_template_name="tt",
+            cmd_args=JaxToolboxCmdArgs(docker_image_url=""),
+        )
 
-    def test_no_profile_stderr_file(self, tmp_path: Path) -> None:
-        result = self.js.get_job_status(tmp_path)
+    def test_no_profile_stderr_file(self, base_tr: TestRun) -> None:
+        result = self.jax_tdef.was_run_successful(base_tr)
         assert not result.is_successful
         assert result.error_message == (
-            f"No profile_stderr_*.txt files found in the specified output directory, {tmp_path}. "
+            f"No profile_stderr_*.txt files found in the specified output directory, {base_tr.output_path}. "
             "These files are expected to be created during the profiling stage. "
             "Please ensure the profiling stage completed successfully. "
             "Run the generated sbatch script manually to debug."
         )
 
-    def test_missing_pax_status_keyword(self, tmp_path: Path) -> None:
-        profile_stderr_file = tmp_path / "profile_stderr_1.txt"
+    def test_missing_pax_status_keyword(self, base_tr: TestRun) -> None:
+        base_tr.output_path.mkdir(parents=True, exist_ok=True)
+        profile_stderr_file = base_tr.output_path / "profile_stderr_1.txt"
         profile_stderr_content = "Some initialization output\nMore output\nFinal output without the expected keyword"
         profile_stderr_file.write_text(profile_stderr_content)
-        result = self.js.get_job_status(tmp_path)
+        result = self.jax_tdef.was_run_successful(base_tr)
         assert not result.is_successful
         assert result.error_message == (
             "The profiling stage completed but did not generate the expected '[PAX STATUS]: E2E time: "
@@ -47,26 +53,28 @@ class TestJaxToolboxJobStatusRetrievalStrategy:
             "You need to run the sbatch script manually to see what happens."
         )
 
-    def test_no_error_files(self, tmp_path: Path) -> None:
-        profile_stderr_file = tmp_path / "profile_stderr.txt"
+    def test_no_error_files(self, base_tr: TestRun) -> None:
+        base_tr.output_path.mkdir(parents=True, exist_ok=True)
+        profile_stderr_file = base_tr.output_path / "profile_stderr.txt"
         profile_stderr_content = "[PAX STATUS]: E2E time: Elapsed time for profiling"
         profile_stderr_file.write_text(profile_stderr_content)
-        result = self.js.get_job_status(tmp_path)
+        result = self.jax_tdef.was_run_successful(base_tr)
         assert not result.is_successful
         assert result.error_message == (
-            f"No profile_stderr_*.txt files found in the specified output directory, {tmp_path}. "
+            f"No profile_stderr_*.txt files found in the specified output directory, {base_tr.output_path}. "
             "These files are expected to be created during the profiling stage. "
             "Please ensure the profiling stage completed successfully. "
             "Run the generated sbatch script manually to debug."
         )
 
-    def test_cuda_no_device_error_in_profile_stderr(self, tmp_path: Path) -> None:
-        profile_stderr_file = tmp_path / "profile_stderr_1.txt"
+    def test_cuda_no_device_error_in_profile_stderr(self, base_tr: TestRun) -> None:
+        base_tr.output_path.mkdir(parents=True, exist_ok=True)
+        profile_stderr_file = base_tr.output_path / "profile_stderr_1.txt"
         profile_stderr_content = "[PAX STATUS]: E2E time: Elapsed time for profiling\n"
         profile_stderr_content += "CUDA_ERROR_NO_DEVICE: no CUDA-capable device is detected"
         profile_stderr_file.write_text(profile_stderr_content)
 
-        result = self.js.get_job_status(tmp_path)
+        result = self.jax_tdef.was_run_successful(base_tr)
         assert not result.is_successful
         assert result.error_message == (
             "CUDA_ERROR_NO_DEVICE: no CUDA-capable device is detected. This may be due to missing "
@@ -78,16 +86,17 @@ class TestJaxToolboxJobStatusRetrievalStrategy:
             "example command."
         )
 
-    def test_missing_e2e_time_keyword(self, tmp_path: Path) -> None:
-        profile_stderr_file = tmp_path / "profile_stderr_1.txt"
+    def test_missing_e2e_time_keyword(self, base_tr: TestRun) -> None:
+        base_tr.output_path.mkdir(parents=True, exist_ok=True)
+        profile_stderr_file = base_tr.output_path / "profile_stderr_1.txt"
         profile_stderr_content = "[PAX STATUS]: E2E time: Elapsed time for profiling"
         profile_stderr_file.write_text(profile_stderr_content)
 
-        error_file = tmp_path / "error-1.txt"
+        error_file = base_tr.output_path / "error-1.txt"
         error_content = "Some initialization output\nMore output\nFinal output without the expected keyword"
         error_file.write_text(error_content)
 
-        result = self.js.get_job_status(tmp_path)
+        result = self.jax_tdef.was_run_successful(base_tr)
         assert not result.is_successful
         assert result.error_message == (
             f"The file {error_file} does not contain the expected 'E2E time: Elapsed time for' keyword at the "
@@ -95,16 +104,17 @@ class TestJaxToolboxJobStatusRetrievalStrategy:
             "Please debug this manually to ensure the actual run stage completes as expected."
         )
 
-    def test_cuda_no_device_error_in_error_file(self, tmp_path: Path) -> None:
-        profile_stderr_file = tmp_path / "profile_stderr_1.txt"
+    def test_cuda_no_device_error_in_error_file(self, base_tr: TestRun) -> None:
+        base_tr.output_path.mkdir(parents=True, exist_ok=True)
+        profile_stderr_file = base_tr.output_path / "profile_stderr_1.txt"
         profile_stderr_content = "[PAX STATUS]: E2E time: Elapsed time for profiling"
         profile_stderr_file.write_text(profile_stderr_content)
 
-        error_file = tmp_path / "error-1.txt"
+        error_file = base_tr.output_path / "error-1.txt"
         error_content = "CUDA_ERROR_NO_DEVICE: no CUDA-capable device is detected"
         error_file.write_text(error_content)
 
-        result = self.js.get_job_status(tmp_path)
+        result = self.jax_tdef.was_run_successful(base_tr)
         assert not result.is_successful
         assert result.error_message == (
             "CUDA_ERROR_NO_DEVICE: no CUDA-capable device is detected. This may be due to missing "
@@ -116,83 +126,89 @@ class TestJaxToolboxJobStatusRetrievalStrategy:
             "example command."
         )
 
-    def test_successful_job(self, tmp_path: Path) -> None:
-        profile_stderr_file = tmp_path / "profile_stderr_1.txt"
+    def test_successful_job(self, base_tr: TestRun) -> None:
+        base_tr.output_path.mkdir(parents=True, exist_ok=True)
+        profile_stderr_file = base_tr.output_path / "profile_stderr_1.txt"
         profile_stderr_content = "[PAX STATUS]: E2E time: Elapsed time for profiling"
         profile_stderr_file.write_text(profile_stderr_content)
 
-        error_file = tmp_path / "error-1.txt"
+        error_file = base_tr.output_path / "error-1.txt"
         error_content = "Some initialization output\nMore output\nE2E time: Elapsed time for actual run\nFinal output"
         error_file.write_text(error_content)
 
-        result = self.js.get_job_status(tmp_path)
+        result = self.jax_tdef.was_run_successful(base_tr)
         assert result.is_successful
         assert result.error_message == ""
 
-    def test_nccl_group_end_error_in_profile_stderr(self, tmp_path: Path) -> None:
-        profile_stderr_file = tmp_path / "profile_stderr_1.txt"
+    def test_nccl_group_end_error_in_profile_stderr(self, base_tr: TestRun) -> None:
+        base_tr.output_path.mkdir(parents=True, exist_ok=True)
+        profile_stderr_file = base_tr.output_path / "profile_stderr_1.txt"
         profile_stderr_content = "[PAX STATUS]: E2E time: Elapsed time for profiling\n"
         profile_stderr_content += "NCCL operation ncclGroupEnd() failed: unhandled system error"
         profile_stderr_file.write_text(profile_stderr_content)
 
-        result = self.js.get_job_status(tmp_path)
+        result = self.jax_tdef.was_run_successful(base_tr)
         assert not result.is_successful
         assert result.error_message == (
             "NCCL operation ncclGroupEnd() failed: unhandled system error. Please check if the NCCL-test "
             "passes. Run with NCCL_DEBUG=INFO for more details."
         )
 
-    def test_nccl_group_end_error_in_error_file(self, tmp_path: Path) -> None:
-        profile_stderr_file = tmp_path / "profile_stderr_1.txt"
+    def test_nccl_group_end_error_in_error_file(self, base_tr: TestRun) -> None:
+        base_tr.output_path.mkdir(parents=True, exist_ok=True)
+        profile_stderr_file = base_tr.output_path / "profile_stderr_1.txt"
         profile_stderr_content = "[PAX STATUS]: E2E time: Elapsed time for profiling"
         profile_stderr_file.write_text(profile_stderr_content)
 
-        error_file = tmp_path / "error-1.txt"
+        error_file = base_tr.output_path / "error-1.txt"
         error_content = "NCCL operation ncclGroupEnd() failed: unhandled system error"
         error_file.write_text(error_content)
 
-        result = self.js.get_job_status(tmp_path)
+        result = self.jax_tdef.was_run_successful(base_tr)
         assert not result.is_successful
         assert result.error_message == (
             "NCCL operation ncclGroupEnd() failed: unhandled system error. Please check if the NCCL-test "
             "passes. Run with NCCL_DEBUG=INFO for more details."
         )
 
-    def test_heartbeat_error_in_profile_stderr(self, tmp_path: Path) -> None:
-        profile_stderr_file = tmp_path / "profile_stderr_1.txt"
+    def test_heartbeat_error_in_profile_stderr(self, base_tr: TestRun) -> None:
+        base_tr.output_path.mkdir(parents=True, exist_ok=True)
+        profile_stderr_file = base_tr.output_path / "profile_stderr_1.txt"
         profile_stderr_content = "[PAX STATUS]: E2E time: Elapsed time for profiling\n"
         profile_stderr_content += "Terminating process because the coordinator detected missing heartbeats"
         profile_stderr_file.write_text(profile_stderr_content)
 
-        result = self.js.get_job_status(tmp_path)
+        result = self.jax_tdef.was_run_successful(base_tr)
         assert not result.is_successful
         assert result.error_message == (
             "Terminating process because the coordinator detected missing heartbeats. This most likely "
             f"indicates that another task died. Please review the file at {profile_stderr_file} and any relevant "
-            f"logs in {tmp_path}. Ensure the servers allocated for this task can reach each other with their "
-            "hostnames, and they can open any ports and reach others' ports."
+            f"logs in {base_tr.output_path}. Ensure the servers allocated for this task can reach each other with "
+            "their hostnames, and they can open any ports and reach others' ports."
         )
 
-    def test_heartbeat_error_in_error_file(self, tmp_path: Path) -> None:
-        profile_stderr_file = tmp_path / "profile_stderr_1.txt"
+    def test_heartbeat_error_in_error_file(self, base_tr: TestRun) -> None:
+        base_tr.output_path.mkdir(parents=True, exist_ok=True)
+        profile_stderr_file = base_tr.output_path / "profile_stderr_1.txt"
         profile_stderr_content = "[PAX STATUS]: E2E time: Elapsed time for profiling"
         profile_stderr_file.write_text(profile_stderr_content)
 
-        error_file = tmp_path / "error-1.txt"
+        error_file = base_tr.output_path / "error-1.txt"
         error_content = "Terminating process because the coordinator detected missing heartbeats"
         error_file.write_text(error_content)
 
-        result = self.js.get_job_status(tmp_path)
+        result = self.jax_tdef.was_run_successful(base_tr)
         assert not result.is_successful
         assert result.error_message == (
             "Terminating process because the coordinator detected missing heartbeats. This most likely "
             f"indicates that another task died. Please review the file at {error_file} and any relevant logs in"
-            f" {tmp_path}. Ensure the servers allocated for this task can reach each other with their "
+            f" {base_tr.output_path}. Ensure the servers allocated for this task can reach each other with their "
             "hostnames, and they can open any ports and reach others' ports."
         )
 
-    def test_pyxis_mktemp_error_in_profile_stderr(self, tmp_path: Path) -> None:
-        profile_stderr_file = tmp_path / "profile_stderr_1.txt"
+    def test_pyxis_mktemp_error_in_profile_stderr(self, base_tr: TestRun) -> None:
+        base_tr.output_path.mkdir(parents=True, exist_ok=True)
+        profile_stderr_file = base_tr.output_path / "profile_stderr_1.txt"
         profile_stderr_content = "[PAX STATUS]: E2E time: Elapsed time for profiling\n"
         profile_stderr_content += (
             "pyxis:     mktemp: failed to create directory via template '/tmp/enroot.XXXXXXXXXX': "
@@ -200,7 +216,7 @@ class TestJaxToolboxJobStatusRetrievalStrategy:
         )
         profile_stderr_file.write_text(profile_stderr_content)
 
-        result = self.js.get_job_status(tmp_path)
+        result = self.jax_tdef.was_run_successful(base_tr)
         assert not result.is_successful
         assert result.error_message == (
             "pyxis: mktemp: failed to create directory via template. This is due to insufficient disk cache "
@@ -210,19 +226,20 @@ class TestJaxToolboxJobStatusRetrievalStrategy:
             "does not have enough disk capacity to cache the Docker image."
         )
 
-    def test_pyxis_mktemp_error_in_error_file(self, tmp_path: Path) -> None:
-        profile_stderr_file = tmp_path / "profile_stderr_1.txt"
+    def test_pyxis_mktemp_error_in_error_file(self, base_tr: TestRun) -> None:
+        base_tr.output_path.mkdir(parents=True, exist_ok=True)
+        profile_stderr_file = base_tr.output_path / "profile_stderr_1.txt"
         profile_stderr_content = "[PAX STATUS]: E2E time: Elapsed time for profiling"
         profile_stderr_file.write_text(profile_stderr_content)
 
-        error_file = tmp_path / "error-1.txt"
+        error_file = base_tr.output_path / "error-1.txt"
         error_content = (
             "pyxis:     mktemp: failed to create directory via template '/tmp/enroot.XXXXXXXXXX': "
             "No space left on device"
         )
         error_file.write_text(error_content)
 
-        result = self.js.get_job_status(tmp_path)
+        result = self.jax_tdef.was_run_successful(base_tr)
         assert not result.is_successful
         assert result.error_message == (
             "pyxis: mktemp: failed to create directory via template. This is due to insufficient disk cache "
