+++ tests/job_status_retrieval_strategy/test_nemo_run_job_status_retrieval_strategy.py
@@ -14,65 +14,73 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-from pathlib import Path
+from cloudai.core import TestRun
+from cloudai.workloads.nemo_run import NeMoRunCmdArgs, NeMoRunTestDefinition
 
-from cloudai.workloads.nemo_run import NeMoRunJobStatusRetrievalStrategy
 
-
-class TestNeMoRunJobStatusRetrievalStrategy:
+class TestNeMoRunSuccessCheck:
     def setup_method(self) -> None:
-        self.js = NeMoRunJobStatusRetrievalStrategy()
+        self.nemorun_tdef = NeMoRunTestDefinition(
+            name="n",
+            description="d",
+            test_template_name="tt",
+            cmd_args=NeMoRunCmdArgs(docker_image_url="", task="", recipe_name=""),
+        )
 
-    def test_no_stderr_file(self, tmp_path: Path) -> None:
-        result = self.js.get_job_status(tmp_path)
+    def test_no_stderr_file(self, base_tr: TestRun) -> None:
+        result = self.nemorun_tdef.was_run_successful(base_tr)
         assert not result.is_successful
         assert result.error_message == (
-            f"stderr.txt file not found in the specified output directory {tmp_path}. "
+            f"stderr.txt file not found in the specified output directory {base_tr.output_path}. "
             "This file is expected to be created as part of the NeMo training job. "
             "Please ensure the job was submitted and executed properly. "
-            f"You can try re-running the job manually and verify that {tmp_path / 'stderr.txt'} is created "
+            f"You can try re-running the job manually and verify that {base_tr.output_path / 'stderr.txt'} is created "
             "with the expected output. If the issue persists, contact the system administrator."
         )
 
-    def test_successful_job(self, tmp_path: Path) -> None:
-        stderr_file = tmp_path / "stderr.txt"
+    def test_successful_job(self, base_tr: TestRun) -> None:
+        base_tr.output_path.mkdir(parents=True, exist_ok=True)
+        stderr_file = base_tr.output_path / "stderr.txt"
         stderr_content = """
         [NeMo Train] Trainer.fit` stopped: `max_steps=100` reached.
         """
         stderr_file.write_text(stderr_content)
-        result = self.js.get_job_status(tmp_path)
+        result = self.nemorun_tdef.was_run_successful(base_tr)
         assert result.is_successful
         assert result.error_message == ""
 
-    def test_missing_max_steps_indicator(self, tmp_path: Path) -> None:
-        stderr_file = tmp_path / "stderr.txt"
+    def test_missing_max_steps_indicator(self, base_tr: TestRun) -> None:
+        base_tr.output_path.mkdir(parents=True, exist_ok=True)
+        stderr_file = base_tr.output_path / "stderr.txt"
         stderr_content = """
         [NeMo Train] Trainer.fit` stopped: training completed successfully.
         """
         stderr_file.write_text(stderr_content)
-        result = self.js.get_job_status(tmp_path)
+        result = self.nemorun_tdef.was_run_successful(base_tr)
         assert not result.is_successful
         assert "'max_steps='" in result.error_message
         assert "'reached'" in result.error_message
 
-    def test_missing_reached_indicator(self, tmp_path: Path) -> None:
-        stderr_file = tmp_path / "stderr.txt"
+    def test_missing_reached_indicator(self, base_tr: TestRun) -> None:
+        base_tr.output_path.mkdir(parents=True, exist_ok=True)
+        stderr_file = base_tr.output_path / "stderr.txt"
         stderr_content = """
         [NeMo Train] Trainer.fit` stopped: `max_steps=100`.
         """
         stderr_file.write_text(stderr_content)
-        result = self.js.get_job_status(tmp_path)
+        result = self.nemorun_tdef.was_run_successful(base_tr)
         assert not result.is_successful
         assert "'reached'" in result.error_message
         assert "'max_steps='" not in result.error_message
 
-    def test_missing_both_indicators(self, tmp_path: Path) -> None:
-        stderr_file = tmp_path / "stderr.txt"
+    def test_missing_both_indicators(self, base_tr: TestRun) -> None:
+        base_tr.output_path.mkdir(parents=True, exist_ok=True)
+        stderr_file = base_tr.output_path / "stderr.txt"
         stderr_content = """
         [NeMo Train] Trainer.fit stopped unexpectedly.
         """
         stderr_file.write_text(stderr_content)
-        result = self.js.get_job_status(tmp_path)
+        result = self.nemorun_tdef.was_run_successful(base_tr)
         assert not result.is_successful
         assert "'max_steps='" in result.error_message
         assert "'reached'" in result.error_message
