+++ tests/test_registry.py
@@ -22,7 +22,6 @@ from cloudai.configurator import BaseAgent
 from cloudai.core import (
     BaseInstaller,
     BaseRunner,
-    JobStatusRetrievalStrategy,
     Registry,
     Reporter,
     ReportGenerationStrategy,
@@ -104,10 +103,6 @@ class AnotherStrategy(TestTemplateStrategy):
     pass
 
 
-class MyJobStatusRetrievalStrategy(JobStatusRetrievalStrategy):
-    pass
-
-
 class TestRegistry__StrategiesMap:
     """This test verifies Registry class functionality.
 
@@ -117,15 +112,8 @@ class TestRegistry__StrategiesMap:
 
     def test_add_strategy(self, registry: Registry):
         registry.add_strategy(MyStrategy, [MySystem], [MyTestDefinition], MyStrategy)
-        registry.add_strategy(
-            MyJobStatusRetrievalStrategy, [MySystem], [MyTestDefinition], MyJobStatusRetrievalStrategy
-        )
 
         assert registry.strategies_map[(MyStrategy, MySystem, MyTestDefinition)] == MyStrategy
-        assert (
-            registry.strategies_map[(MyJobStatusRetrievalStrategy, MySystem, MyTestDefinition)]
-            == MyJobStatusRetrievalStrategy
-        )
 
     def test_add_strategy_duplicate(self, registry: Registry):
         with pytest.raises(ValueError) as exc_info:
